+++ src/test/java/baseball/domain/BallTest.java
@@ -1,5 +1,7 @@
-package baseball;
+package baseball.domain;
 
+import baseball.domain.Ball;
+import baseball.domain.BallStatus;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
